+++ src/context.js
@@ -1,5 +1,6 @@
 const context = {
   argv: null,
+  ext: null,
 }
 
 module.exports = context